Numbers Exercise
Here I am working with numbers and working through this programming exercise.

(Count positive and negative numbers and compute the average of numbers)
Write a program that reads an specified number of integers, determines how many positive and negative values have been read, and computes the total and average of
the input values. Display the average.